Merge rmtheis:master

+++ tess-two/jni/com_googlecode_leptonica_android/src/src/zlibmem.c
@@ -132,21 +132,23 @@ z_stream  z;
             z.next_in = bufferin;
             bbufferWrite(bbin, bufferin, L_BUF_SIZE, &nbytes);
 #if DEBUG
-            fprintf(stderr, " wrote %d bytes to bufferin\n", nbytes);
+            fprintf(stderr, " wrote %lu bytes to bufferin\n",
+                    (unsigned long)nbytes);
 #endif  /* DEBUG */
             z.avail_in = nbytes;
         }
         flush = (bbin->n) ? Z_SYNC_FLUSH : Z_FINISH;
         status = deflate(&z, flush);
 #if DEBUG
-        fprintf(stderr, " status is %d, bytesleft = %d, totalout = %d\n",
+        fprintf(stderr, " status is %d, bytesleft = %u, totalout = %lu\n",
                   status, z.avail_out, z.total_out);
 #endif  /* DEBUG */
         nbytes = L_BUF_SIZE - z.avail_out;
         if (nbytes) {
             bbufferRead(bbout, bufferout, nbytes);
 #if DEBUG
-            fprintf(stderr, " read %d bytes from bufferout\n", nbytes);
+            fprintf(stderr, " read %lu bytes from bufferout\n",
+                    (unsigned long)nbytes);
 #endif  /* DEBUG */
         }
         z.next_out = bufferout;
